# docs/swagger.yaml
basePath: /api
definitions:
...
  im_messages.ImMessages:
    properties:
      created_at:
        type: string
      data:
        type: string
      form_id:
        type: integer
      id:
        type: integer
      is_read:
        type: integer
      msg:
        type: string
      msg_type:
        type: integer
      status:
        type: integer
      to_id:
        type: integer
      users:
        $ref: '#/definitions/im_messages.ImUsers'
    type: object
  im_messages.ImUsers:
    properties:
      avatar:
        description: 头像
        type: string
      email:
        type: string
      id:
        type: integer
      name:
        type: string
    type: object
  im_sessions.ImSessions:
    properties:
      avatar:
        description: 会话头像
        type: string
      channel_type:
        description: 0.单聊 1.群聊
        type: integer
      created_at:
        type: string
      form_id:
        type: integer
      id:
        description: 会话表
        type: integer
      name:
        description: 会话名称
        type: string
      note:
        description: 备注
        type: string
      status:
        description: 会话状态 0.正常 1.禁用
        type: integer
      to_id:
        type: integer
      top_status:
        description: 0.否 1.是
        type: integer
      top_time:
        type: string
      users:
        $ref: '#/definitions/im_sessions.ImUsers'
    type: object
  im_sessions.ImUsers:
    properties:
      age:
        type: integer
      avatar:
        description: 头像
        type: string
      bio:
        description: 用户简介
        type: string
      client_type:
        description: 1.web 2.pc 3.app
        type: integer
      email:
        type: string
      id:
        description: model.BaseModel
        type: integer
      last_login_time:
        description: 最后登录时间
        type: string
      name:
        type: string
      sex:
        description: 0 未知 1.男 2.女
        type: integer
      status:
        description: 0 离线 1 在线
        type: integer
    type: object
...
  response.JsonResponse:
    properties:
      code:
        description: 状态码
        type: integer
      data:
        description: 返回结构数据
      httpCode:
        description: http状态码
        type: integer
      message:
        description: 消息内容
        type: string
    type: object
...
host: 127.0.0.1:8000
info:
  contact:
    email: pltrueover@gmail.com
    name: im-services
  license:
    name: MIT
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: im-services  接口文档
  version: "1.0"
paths:
...
  /messages/:
    get:
      description: 获取私聊消息列表
      parameters:
      - description: 'Bearer '
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/im_messages.ImMessages'
                  type: array
              type: object
      summary: friends/record 获取私聊消息列表
      tags:
      - 消息
  /messages/private:
    post:
      description: 发送私聊消息
      parameters:
      - description: 'Bearer '
        in: header
        name: Authorization
        required: true
        type: string
      - description: 客户端消息id
        in: formData
        name: msg_client_id
        required: true
        type: integer
      - description: 推送人id
        in: formData
        name: to_id
        required: true
        type: integer
      - description: 消息类型:1.私聊消息 2.图片消息 3.语音消息 ..
        in: formData
        name: msg_type
        required: true
        type: integer
      - description: 消息内容
        in: formData
        name: message
        required: true
        type: string
      - description: 自定义携带消息
        in: formData
        name: data
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/im_messages.ImMessages'
                  type: array
              type: object
      summary: friends/record 发送私聊消息
      tags:
      - 消息
  /sessions/:
    get:
      description: 获取会话列表
      parameters:
      - description: 'Bearer '
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/im_sessions.ImSessions'
                  type: array
              type: object
      summary: sessions/ 获取会话列表
      tags:
      - 会话
    post:
      description: 添加会话
      parameters:
      - description: 'Bearer '
        in: header
        name: Authorization
        required: true
        type: string
      - description: 会话类型
        in: formData
        name: type
        required: true
        type: integer
      - description: 聊天对象id
        in: formData
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            allOf:
            - $ref: '#/definitions/response.JsonResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/im_sessions.ImSessions'
                  type: array
              type: object
      summary: sessions/ 添加会话
      tags:
      - 会话
  /sessions/:id:
    delete:
      description: 删除会话
      parameters:
      - description: 'Bearer '
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/response.JsonResponse'
      summary: sessions/:id 删除会话
      tags:
      - 会话
    put:
      description: 更新会话
      parameters:
      - description: 'Bearer '
        in: header
        name: Authorization
        required: true
        type: string
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      - description: 置顶 0 取消 1置顶
        in: formData
        name: top_status
        required: true
        type: integer
      - description: 会话备注
        in: formData
        name: note
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: ok
          schema:
            $ref: '#/definitions/response.JsonResponse'
      summary: sessions/:id 更新会话
      tags:
      - 会话
...
swagger: "2.0"